+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Cannot modify CVE of existing flaws (OSIDB-3102)
+- Jira metadata collector is not deleting metadata on failure (OSIDB-3219)
 
 ## [4.1.6] - 2024-08-02
 ### Fixed